Moving a wrapper for deprecated sdf API moved to a separate class
