+++ rotors_gazebo_plugins/include/rotors_gazebo_plugins/common.h
@@ -106,29 +106,6 @@ discretized system (ZoH):
     T previousState_;
 };
 
-class SdfVector3 : public sdf::Vector3 {
-/*
-A wrapper class for deprecated sdf::Vector3 class to provide the same accessor
-functions as in the newer ignition::math::Vector3 class.
-*/
-
-  public:
-    using sdf::Vector3::Vector3;
-    virtual ~SdfVector3() {}
-
-    /// \brief Get the x value
-    /// \return the x value
-    double X() { return this->x; }
-
-    /// \brief Get the y value
-    /// \return the y value
-    double Y() { return this->y; }
-
-    /// \brief Get the z value
-    /// \return the z value
-    double Z() { return this->z; }
-};
-
 /// Computes a quaternion from the 3-element small angle approximation theta.
 template<class Derived>
 Eigen::Quaternion<typename Derived::Scalar> QuaternionFromSmallAngle(const Eigen::MatrixBase<Derived> & theta) {
